只有登入的使用者可以對自己的留言進行修改刪除

+++ resources/views/posts/show.blade.php
@@ -127,9 +127,11 @@
 
                                     <div class="comment-info">
                                         <div class="comment-author">
-                                            <a href="#">{{ $comment->user->name }}</a>
-                                            <button class="btn btn-default" onclick="toggleCommentForm(event)">編輯</button>
-                                            <button class="btn btn-default" onclick="deleteComment(event)" data-action="/comments/{{ $comment->id }}">刪除</button>
+                                            <a href="#">{{ $comment->name }}</a>
+                                            @if($comment->user && $comment->user->id == Auth::id())
+                                                <button class="btn btn-default" onclick="toggleCommentForm(event)">編輯</button>
+                                                <button class="btn btn-default" onclick="deleteComment(event)" data-action="/comments/{{ $comment->id }}">刪除</button>
+                                            @endif
                                         </div>
                                         {{ $comment->created_at->format('F d, Y').', at '.$comment->created_at->format('G:i') }}
                                     </div>
